attempt to make editor contents consistent across sessions
added FULL_TEXT change type, for full text of editor window.
clicking submit results in sending a full text change,
waiting until the server has acknowledged receiving and
persisting it, and then doing the submission and testing.
the editor window is read-only while a submission is in progress.

extended LoadExerciseService to allow retrieval of user's
latest full text for problem.  if found, editor is populated
with this text.

Note: need to change logout so that a fulltext change is sent.
Otherwise, editor contents won't be accurate for next session.