changed RPC services to perform authentication
SubmitServiceImpl still needs to be changed.

The idea is that we always have to check to make sure that actions
are performed only if the client is authenticated and has the right
to perform the action.  We have to avoid leaking information that
the client shouldn't have access to, forging data, etc.